requirements.txt
Django==1.7.1
django-allauth==0.18.0